djangoapi/requirements.txt
Django==2.2
django-filter==2.1.0
djangorestframework==3.9.2
Markdown==3.1
mysqlclient==1.4.2.post1
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
